Issue #19 fixed
