--- app/src/main/java/com/coderbunker/kioskapp/KioskWebviewClient.java
@@ -1,24 +1,0 @@
-package com.coderbunker.kioskapp;
-
-import android.net.http.SslError;
-import android.webkit.SslErrorHandler;
-import android.webkit.WebView;
-import android.webkit.WebViewClient;
-import android.widget.Toast;
-
-class KioskWebviewClient extends WebViewClient {
-
-    @Override
-    public boolean shouldOverrideUrlLoading(WebView view, String url) {
-        if (url.contains(url)) {
-            view.loadUrl(url);
-        }
-        return true;
-
-    }
-
-    @Override
-    public void onReceivedSslError(WebView view, SslErrorHandler handler, SslError error) {
-        handler.proceed(); //Ignore SSL certificate error
-    }
-}
